src/get_token/requirements.txt
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
pyperclip

